ref(shared): move PhelFunction and CompletionResultTransfer into Shared
ApiFacadeInterface (in Shared) referenced Api\Transfer DTOs, creating a
Shared->Api dependency cycle that chained Build, Command and Interop into one
large strongly connected component. Both transfers are dependency-free DTOs
consumed by Api, Lsp, Nrepl and Run; they now live in Shared\Api.
