+++ src/php/Api/Application/PhelFnNormalizer.php
@@ -7,9 +7,9 @@ namespace Phel\Api\Application;
 use Phel\Api\Domain\PhelFnGroupKeyGeneratorInterface;
 use Phel\Api\Domain\PhelFnLoaderInterface;
 use Phel\Api\Domain\PhelFnNormalizerInterface;
-use Phel\Api\Transfer\PhelFunction;
 use Phel\Lang\Collections\Map\PersistentMapInterface;
 use Phel\Lang\Keyword;
+use Phel\Shared\Api\PhelFunction;
 use Phel\Shared\ScalarCoercion;
 
 final readonly class PhelFnNormalizer implements PhelFnNormalizerInterface
