+++ tests/php/Integration/Api/ApiFacadeTest.php
@@ -7,9 +7,9 @@ namespace PhelTest\Integration\Api;
 use Phel;
 use Phel\Api\ApiConfig;
 use Phel\Api\ApiFacade;
-use Phel\Api\Transfer\PhelFunction;
 use Phel\Compiler\Infrastructure\GlobalEnvironmentSingleton;
 use Phel\Lang\Symbol;
+use Phel\Shared\Api\PhelFunction;
 use PHPUnit\Framework\Attributes\PreserveGlobalState;
 use PHPUnit\Framework\Attributes\RunInSeparateProcess;
 use PHPUnit\Framework\TestCase;
